+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * [#1546](https://github.com/oshi/oshi/pull/1546): Expose network interface operational status (Windows and Linux) - [@dornand](https://github.com/dornand).
 * [#1548](https://github.com/oshi/oshi/pull/1548): Add getter for descendants of a process - [@dbwiddis](https://github.com/dbwiddis).
 
+##### Bug fixes / Improvements
+* [#1552](https://github.com/oshi/oshi/pull/1552): Handle exceptions querying Windows user info - [@dbwiddis](https://github.com/dbwiddis).
+
 # 5.5.0 (2021-02-08), 5.5.1 (2021-02-21)
 
 ##### New Features